refactor(nx-cmake): filterGccDependencyOutput, filterProjects
