+++ packages/nx-cmake/src/utils/graphUtils/filterGccDependencyOutput/filterGccDependencyOutput.ts
@@ -1,16 +1,4 @@
-export const isValidGccOutput = (
-    chunk: string,
-    originalFile: string
-): boolean => {
-    return (
-        /\.(h|c|cpp)/.test(chunk) &&
-        chunk !== originalFile &&
-        !chunk.includes('.o') &&
-        !chunk.startsWith('include/') &&
-        !chunk.startsWith('/usr/') &&
-        !chunk.startsWith('dist/')
-    );
-};
+import { isValidGccOutput } from './isValidGccOutput/isValidGccOutput';
 
 export const filterGccDependencyOutput = (
     output: string,
