+++ packages/nx-cmake/src/utils/graphUtils/filterGccDependencyOutput/isValidGccOutput/isValidGccOutput.spec.ts
@@ -1,0 +1,3 @@
+describe('isValidGccOutput', () => {
+    it.todo('should test if chunk is valid gcc output');
+});
