+++ packages/nx-cmake/src/utils/graphUtils/filterProjects/filterTags/filterTags.ts
@@ -1,0 +1,3 @@
+import { isC } from '../isC/isC';
+
+export const filterTags = (tags: string[]) => tags.filter(isC);
