+++ packages/nx-cmake/src/utils/graphUtils/filterProjects/getTag/getTag.ts
@@ -1,0 +1,4 @@
+import type { CTag } from '../../../../models/types';
+import { isC } from '../isC/isC';
+
+export const getTag = (tags: string[]): CTag => tags.find<CTag>(isC);
