+++ packages/nx-cmake/src/utils/graphUtils/filterProjects/isC/isC.spec.ts
@@ -1,0 +1,13 @@
+import { isC } from './isC';
+
+describe('isC', () => {
+    it('should return true for "c" or "cpp"', () => {
+        expect(isC('c')).toBe(true);
+        expect(isC('cpp')).toBe(true);
+    });
+
+    it('should return false for other strings', () => {
+        expect(isC('java')).toBe(false);
+        expect(isC('python')).toBe(false);
+    });
+});
