+++ packages/nx-cmake/src/utils/graphUtils/filterProjects/isC/isC.ts
@@ -1,0 +1,3 @@
+import type { CTag } from '../../../../models/types';
+
+export const isC = (s: string): s is CTag => s === 'c' || s === 'cpp';
